create adaptor for std::tuple type
